+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Improve agent management action descriptions for clarity
 ---
 
 ## [0.40.0] - 2026-04-14